Add files via upload
Question 2
A hospital wants to determine whether there is any difference in the average Turn Around Time (TAT) of reports of the laboratories on their preferred list. They collected a random sample and recorded TAT for reports of 4 laboratories. TAT is defined as a sample collected to report dispatch.
Analyze the data and determine whether there is any difference in average TAT among the different laboratories at a 5% significance level.
Dataset: LabTAT.csv