Add files via upload
Question-1
An F&B manager wants to determine whether there is any significant difference in the diameter of the cutlet between the two units. A randomly selected sample of cutlets was collected from both units and measured? Analyze the data and draw inferences at 5% significance level. Please state the assumptions and tests that you carried out to check the validity of the assumptions.
Dataset: Cutlets.csv